+++ examples/data-sources/nxos_ethernet_settings/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_ethernet_settings" "example" {
+}