+++ examples/resources/nxos_ethernet_settings/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_ethernet_settings.example "sys/ethpm/inst"