+++ internal/provider/provider.go
@@ -241,6 +241,7 @@ func (p *provider) GetResources(ctx context.Context) (map[string]tfsdk.ResourceT
 		"nxos_subinterface_vrf":       resourceSubinterfaceVRFType{},
 		"nxos_dhcp_relay_interface":   resourceDHCPRelayInterfaceType{},
 		"nxos_dhcp_relay_address":     resourceDHCPRelayAddressType{},
+		"nxos_ethernet_settings":      resourceEthernetSettingsType{},
 	}, nil
 }
 
@@ -267,6 +268,7 @@ func (p *provider) GetDataSources(ctx context.Context) (map[string]tfsdk.DataSou
 		"nxos_subinterface_vrf":       dataSourceSubinterfaceVRFType{},
 		"nxos_dhcp_relay_interface":   dataSourceDHCPRelayInterfaceType{},
 		"nxos_dhcp_relay_address":     dataSourceDHCPRelayAddressType{},
+		"nxos_ethernet_settings":      dataSourceEthernetSettingsType{},
 	}, nil
 }